Fix up font path and container width
For some reason the requested fonts were producing errors so get new
Google font links for the fonts that were requested.  But Droid Sans
is no longer available so use Open Sans instead.

And finally make the container width a little larger so that the Demo
button 5 isn't wrapped to the next "line".  It's all on one line now,
even with the default fallback Arial fonts.