Disable unneeded sliders and do some cleanup
Disable sliders for the parameters that aren't used for the selected
filter type.  For example, the gain parameter isn't used for
lowpass/highpass filters, so disable the slider.

Also adjust the plot axis limits based on the limits from the slider
for some filter types.  Useful for lowshelf and highshelf so that the
gain parameter doesn't cause the mag response to be clipped.